petcare-backend/requirements.txt
django==4.2.29
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
